fixes: #67 getConfigSource as sorted array.
This is needed as most use cases need the ConfigSources in the effectively sorted order.
They also need it sometimes in descending and for other use cases in ascending ordinal order.
This cannot be achieved by an Iterable in easy ways.
Array has the implicit advantage that it is automatically immutable.

Signed-off-by: Mark Struberg <struberg@apache.org>